+++ browser_use/dom/views.py
@@ -358,7 +358,7 @@ class SerializedDOMState:
 		include_attributes: list[str] | None = None,
 	) -> str:
 		"""Kinda ugly, but leaving this as an internal method because include_attributes are a parameter on the agent, so we need to leave it as a 2 step process"""
-		from browser_use.dom.serializer import DOMTreeSerializer
+		from browser_use.dom.serializer.serializer import DOMTreeSerializer
 
 		if not self._root:
 			return 'Empty DOM tree'
